refactor: drop newly added tests
